refactor(st-09086): modularize schema diff utilities
